requirements.txt
python-dotenv==1.0.1
python-telegram-bot==22.6
pytz==2025.1
requests==2.32.3
beautifulsoup4==4.13.3
lxml==5.3.1
